refactor: 커뮤니티 페이지
