+++ app/src/main/java/com/kau/ttokttok/ui/compose/community/detail/CommunityDetailFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
+import androidx.navigation.findNavController
 import androidx.navigation.fragment.findNavController
 
 class CommunityDetailFragment : Fragment() {
@@ -21,7 +22,12 @@ class CommunityDetailFragment : Fragment() {
 
         setContent {
             CommunityDetailRoute(
-                onClickBack = { findNavController().popBackStack() }
+                onClickBack = {
+                    findNavController().previousBackStackEntry?.
+                    savedStateHandle?.set("needRefresh", true)
+
+                    findNavController().popBackStack()
+                }
             )
         }
     }
