+++ app/src/main/java/com/kau/ttokttok/ui/compose/preconsideration/writing/WritingPreConsiderationScreen.kt
@@ -27,10 +27,10 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.kau.ttokttok.ui.component.common.button.PostBottomActions
 import com.kau.ttokttok.ui.component.common.header.WhiteHeader
-import com.kau.ttokttok.ui.compose.community.writing.PostBottomActions
-import com.kau.ttokttok.ui.compose.community.writing.PostContentField
-import com.kau.ttokttok.ui.compose.community.writing.PostTitleField
+import com.kau.ttokttok.ui.component.common.textfield.PostContentField
+import com.kau.ttokttok.ui.component.common.textfield.PostTitleField
 import com.kau.ttokttok.ui.theme.*
 
 @Preview
